NF(GIFT for Michael): obj.states.find('sensitivities')
Now you can search through the container for what could have an
interesting for you state. It is not yet tested in real life, but
works on a test case. The argument is a regular expression for
python's re so could be quite flexible. Now I hope we will not get
lost

TODO: make it work for iterable members of the instance (just recalled
about it... but better first clarify with Michael if that is worth
doing at all)