NF+BF: Fix handling of empty feature set in OptimalOpverlapThresholds
So far a transfer error of 1.0 was assumed when no feature was present in
some feature set (overlap, spread, never-selected). This is/was stupid.

Now this case is handled with masked array. Fill values for masked array are
set to large negative value (-1000) to make sure that any bug in the script
will really corrupt any result and make the bug obvious -- I hope ;-)

Also new feature is that the class now uses the inner transfer error of the
spread feature to determine optimal threshold. First time the error drops
below some critical value overlapping voxels at that stage are used perform
feature selection on original dataset(s).